requirements.txt
asgiref==3.8.1
Django==3.2.7
django-ratelimit==4.1.0
djangorestframework==3.14.0
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.12.2
gunicorn==20.1.0